Here are some things you should have thought about:

**Multi-Tenancy**
- [ ] Extended new entities with `AbstractTenantAwareEntity`?
- [ ] New entity added to `TenantAwareDatabaseConfiguration`?
- [ ] Tested with `dev-multitenant` profile?

<!--

Thanks for contributing to the zeiterfassung.
Please review the following notes before submitting you pull request.

Please look for other issues or pull requests which already work on this
topic. Is somebody already on it? Do you need to synchronize?

# Security Vulnerabilities

🛑 STOP! 🛑 If your contribution fixes a security vulnerability, please do
not submit it.
Instead, please write an E-Mail to info@focus-shift.de with all the
information
to recreate the security vulnerability.

# Describing Your Changes

If, having reviewed the notes above, you're ready to submit your pull
request, please
provide a brief description of the proposed changes.

If they:
🐞 fix a bug, please describe the broken behaviour and how the changes
fix it.
    Please label with 'type: bug' and 'status: new'
    
🎁 make an enhancement, please describe the new functionality and why you
believe it's useful.
    Please label with 'type: enhancement' and 'status: new'
 
If your pull request relates to any existing issues,
please reference them by using the issue number prefixed with #.

-->

+++ src/main/java/de/focusshift/zeiterfassung/integration/urlaubsverwaltung/application/ApplicationEventDtoAdapter.java
@@ -16,7 +16,7 @@ import java.util.Set;
 class ApplicationEventDtoAdapter {
 
     private final String tenantId;
-    private final Integer sourceId;
+    private final Long sourceId;
     private final ApplicationPersonDTO person;
     private final VacationTypeDTO vacationType;
     private final ApplicationPeriodDTO period;
@@ -53,7 +53,7 @@ class ApplicationEventDtoAdapter {
         return tenantId;
     }
 
-    public Integer getSourceId() {
+    public Long getSourceId() {
         return sourceId;
     }
